requirements.txt
requests>=2.31.0
websockets==12.0
influxdb-client>=1.36.0
python-dotenv>=1.0.0
jwt==1.3.1
pandas>=2.0.0
aiohttp==3.9.1
numpy>=1.24.0
torch>=2.0.0
gym>=0.26.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
hmmlearn>=0.3.0
websocket-client>=1.5.0
python-dateutil>=2.8.2 